+++ packages/ddp-client/livedata_tests.js
@@ -715,8 +715,9 @@ if (Meteor.isClient) {
 
   testAsyncMulti("livedata - publish multiple cursors", [
     function (test, expect) {
-      Meteor.subscribe("multiPublish", {normal: 1}, {
+      var sub = Meteor.subscribe("multiPublish", {normal: 1}, {
         onReady: expect(function () {
+          test.isTrue(sub.ready());
           test.equal(One.find().count(), 2);
           test.equal(Two.find().count(), 3);
         }),