+++ tools/main-transpile-wrapper.js
@@ -1,7 +1,15 @@
 // This file exists because it is the file in the tool that is not automatically
 // transpiled by Babel
 
-require('meteor-babel/register'); // #RemoveInProd this line is removed in isopack.js
+function babelRegister() {
+  require("meteor-babel/register")({
+    babelOptions: require("meteor-babel").getDefaultOptions(
+      require("./babel-features.js")
+    )
+  });
+}
+
+babelRegister(); // #RemoveInProd this line is removed in isopack.js
 
 // Install a global ES6-compliant Promise constructor that knows how to
 // run all its callbacks in Fibers.